requirements.txt
pyzmq>=25,<27
pyserial>=3.5,<4
websockets>=12,<15
